+++ CHANGELOG.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.0] - 2024-07-23
+## [Unreleased](https://github.com/deepgram-devs/deepgram-rust-sdk/compare/0.6.0...HEAD)
+
+## [0.6.0](https://github.com/deepgram-devs/deepgram-rust-sdk/compare/0.5.0...0.6.0) - 2024-08-08
 
 ### Migrating from 0.4.0 -> 0.6.0
 
-Module Imports
+#### Module Imports
 
-```
+```rust
 use deepgram::{
 ---    transcription::prerecorded::{
 +++    common::{
@@ -21,10 +23,49 @@ use deepgram::{
 };
 ```
 
-Streaming Changes
+#### Streaming Changes
+
+We have exposed a low-level, message-based interface to the websocket API:
+
+```rust
+use futures::select;
+
+let mut handle = dg
+    .transcription()
+    .stream_request()
+    .handle()
+    .await?;
+
+loop {
+    select! {
+        _ = tokio::time::sleep(Duration::from_secs(3)) => handle.keep_alive().await,
+        _ = handle.send_data(data_chunk()).fuse() => {}
+        response = handle.receive().fuse() => {
+            match response {
+                Some(response) => println!("{response:?}"),
+                None => break,
+            }
+        }
+    }
+}
+handle.close_stream().await;
+```
+
+No need to call `.start()` to begin streaming data.
+
+```rust
+let mut results = dg
+    .transcription()
+    .stream_request_with_options(Some(&options))
+    .file(PATH_TO_FILE, AUDIO_CHUNK_SIZE, Duration::from_millis(16))
+---    .await
+---    .start()
+    .await;
+```
 
 Now you can pass Options using stream_request_with_options
-```
+
+```rust
 let options = Options::builder()
     .smart_format(true)
     .language(Language::en_US)
@@ -35,8 +76,6 @@ let mut results = dg
     .stream_request_with_options(Some(&options))
     .file(PATH_TO_FILE, AUDIO_CHUNK_SIZE, Duration::from_millis(16))
     .await?
-    .start()
-    .await?;
 ```
 
 Some Enums have changed and may need to be updated
@@ -47,6 +86,7 @@ Some Enums have changed and may need to be updated
 - Add support for pre-recorded features when streaming
 - Add Speech to Text
 - Reorganize Code
+
 
 ### Streaming Features
 - endpointing
@@ -86,13 +126,14 @@ Some Enums have changed and may need to be updated
 - custom_topics
 - custom_topic_mode
 
-## [0.5.0]
+## [0.5.0](https://github.com/deepgram-devs/deepgram-rust-sdk/compare/0.4.0...0.5.0) - 2024-07-08
+
 - Deprecate tiers and add explicit support for all currently available models.
 - Expand language enum to include all currently-supported languages.
 - Add (default on) feature flags for live and prerecorded transcription.
 - Support arbitrary query params in transcription options.
 
-## [0.4.0] - 2023-11-01
+## [0.4.0](https://github.com/deepgram-devs/deepgram-rust-sdk/compare/0.3.0...0.4.0) - 2023-11-01
 
 ### Added
 - `detect_language` option.
@@ -101,7 +142,7 @@ Some Enums have changed and may need to be updated
 - Remove generic from `Deepgram` struct.
 - Upgrade dependencies: `tungstenite`, `tokio-tungstenite`, `reqwest`.
 
-## [0.3.0]
+## [0.3.0](https://github.com/deepgram-devs/deepgram-rust-sdk/compare/0.2.1...0.3.0) - 2023-07-26
 
 ### Added
 - Derive `Serialize` for all response types.
@@ -113,6 +154,3 @@ Some Enums have changed and may need to be updated
 ### Changed
 - Use Rustls instead of OpenSSL.
 
-[Unreleased]: https://github.com/deepgram-devs/deepgram-rust-sdk/compare/0.4.0...HEAD
-[0.4.0]: https://github.com/deepgram-devs/deepgram-rust-sdk/compare/0.3.0...0.4.0
-[0.3.0]: https://github.com/deepgram-devs/deepgram-rust-sdk/compare/0.2.1...0.3.0